[Editorial] Edit case of top-level single-quotes.
This is a bit tricky. Lint dislikes the use of single-quotes here, but
it doesn't make sense to just flatly change them to double-quotes.
From context (the use of third-person in particular), it's clear that
the single-quoted text is not meant to be read literally, but is
instead generally representative of what the old man says.

My feeling is that the best way to fix this is to leave out the
quotes, as this is how reported dialog that has been reframed to
convey the sense rather than the actual words is usually written. The
construct should be familiar to most readers.

Unfortunately, the first bit of single-quoted "dialog" is light on
punctuation, and taking out the quote marks as well makes hash of the
sentence. So I've retained the quotes on that sentence (and making
them double-quotes, so as to pacify lint), as this sentence still
makes perfect sense when the dialog is read literally.