[Editorial] Insert various commas so that lint will stop complaining.
Dickens often omits commas around dialog and dialog tags (at least in
this edition), so I'm not sure why these two places are being singled
out by lint. But they are, so here we go.

Note also that one of these two places was the spot where quotes were
retained solely because of a lack of punctuation. Since lint is still
unhappy, even with the double-quotes, clearly the way forward is to
add the commas and then remove the single-quote marks, as originally
planned.